Refactor reporting classses to uniform behaviours
